move timer to separate file
